generic_nvram - Remove calls to whoop init & exit
The return type for nvram_init() does not match the signature expected
by the _whoop_init function pointer, so SMACK doesn't match the function
and analysis stops.

More generally (for svcomp purposes, that is), we shouldn't be trying to
assign a function's address to a function pointer when the signatures
don't match, anyway.

Perhaps I'll readress this at some point, and try to figure out how Akis
is using these _whoop_init and _whoop_exit function pointers, but for
now, I'll just not use these