Project - add list of entrypoints and races
This xml file lists entry points from each driver that can be called
concurrently.  For each pair of entry points, it also lists whether
Whoop detected races with that pair of entry points.

This file is to be used to determine for which pairs of entry points a
benchmark should be created, and whether that pair is thought to have a
race.