pthreadify - flip whether bug is expected in filename
entry_points.xml, provided by Akis, lists whether or not a bug is
expected.  svcomp benchmarks should be marked as true if there is no bug,
and false if a bug is known to exist in the benchmark.

I initially overlooked this, and so the output filename for each entry
point pair was the opposite of what it should be.