pthreadify - further correcting buggy indicator
I changed this and myChauffeur so that the bugginess indicator is called
"noBug", and is true when noBug exists (that is, file is free from bugs).

This is consistent with the svcomp naming conventions.

At this point, the only place that the original "bug" indicator exists
is in entry_points.xml