generic_nvram - update after fixing myChauffeur
myChauffeur had to be updated because the AST structure changed after
the signature of the function pointers for init functions actually
matched the signature of the functions being pointed to (prior to the
fix, there was an additional BitCast node).

This commit regenerates the generic_nvram benchmarks after applying
this fix to myChauffeur.