pthreadify - ignore pairs where entry point 1 == 2
In order to avoid having to research how each driver type should be
called when the entry point is being called concurrently with itself,
we decided to only include entry point pairs where entry point 1 is not
the same as entry point 2.

This commit does this, and also deletes such benchmarks for the
generic_nvram driver.